workload/image_classification/src/subscriber_serverless/function/requirements.txt
numpy
Pillow==8.1.0